+++ frontend/src/screens/CheckCalendar.tsx
@@ -1,5 +1,6 @@
 import {
   Keyboard,
+  SafeAreaView,
   StyleSheet,
   Text,
   TouchableOpacity,
@@ -17,35 +18,42 @@ import { useRecoilValue } from 'recoil';
 import { Calendar } from 'react-native-calendars';
 import MonthPicker from 'react-native-month-year-picker';
 import SelectDropdown from 'react-native-select-dropdown';
-import { userState } from '../../state/atoms/userAtom';
-import { getFormattedDate } from '../../utils/formattedDate';
-import Plus from '../../assets/images/common/plus.svg';
-import CheckCalendarDayComponent from './CheckCalendarDayComponent';
+import { StackScreenProps } from '@react-navigation/stack';
+import { userState } from '../state/atoms/userAtom';
+import { getFormattedDate } from '../utils/formattedDate';
+import Plus from '../assets/images/common/plus.svg';
+import CheckCalendarDayComponent from '../components/plan/CheckCalendarDayComponent';
 import {
   DayType,
   PlanPropsType,
   SchedulesType,
-} from '../../types/calendar/calendarType';
-import { handleGetPlan, handleGetPlanDetail } from '../../api/plan/getPlan';
-import { partnerState } from '../../state/atoms/partnerAtom';
-import OptionArrow from '../../assets/images/common/optionArrow.svg';
-import CheckCalendarDetail from './CheckCalendarDetail';
-import { month } from '../../utils/plan/calendarText';
+} from '../types/calendar/calendarType';
+import { handleGetPlan, handleGetPlanDetail } from '../api/plan/getPlan';
+import { partnerState } from '../state/atoms/partnerAtom';
+import OptionArrow from '../assets/images/common/optionArrow.svg';
+import CheckCalendarDetail from '../components/plan/CheckCalendarDetail';
+import { month } from '../utils/plan/calendarText';
 import {
   handleCheckAnniversaryList,
   handleCheckCouplePlanDetail,
   handleCheckMyPlanDetail,
   handleCheckPartnerPlanDetail,
-} from '../../utils/plan/calendar';
-import { coupleState } from '../../state/atoms/coupleAtom';
+} from '../utils/plan/calendar';
+import { coupleState } from '../state/atoms/coupleAtom';
+import { StackParamList } from '../types/routes/navigationType';
 
-const CheckCalendar = ({
-  detail,
-  setDaySelected,
-}: {
-  detail?: boolean;
-  setDaySelected?: Dispatch<SetStateAction<string>>;
-}) => {
+type Props = StackScreenProps<StackParamList, 'PlanCalendarScreen'>;
+
+const CheckCalendar = (
+  { navigation }: Props,
+  {
+    detail,
+    setDaySelected,
+  }: {
+    detail?: boolean;
+    setDaySelected?: Dispatch<SetStateAction<string>>;
+  }
+) => {
   const userData = useRecoilValue(userState);
   const partnerData = useRecoilValue(partnerState);
   const coupleData = useRecoilValue(coupleState);
@@ -148,6 +156,11 @@ const CheckCalendar = ({
     };
   }
 
+  // 일정 추가 페이지로 넘어가기
+  const handleAddPlan = (id: number | null) => {
+    navigation.navigate('PlanRoute', { id });
+  };
+
   useEffect(() => {
     const year = selectedYear || currentYear;
     const month = selectedMonth || currentMonth;
@@ -206,129 +219,136 @@ const CheckCalendar = ({
 
   return (
     <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
-      <View style={marginContainerStyle}>
-        <View style={detail ? undefined : styles.calendarView}>
-          {/* 플랜 생성 페이지 분기처리 */}
-          {detail ? (
-            <View style={styles.calendarHeaderView}>
-              <TouchableOpacity
-                activeOpacity={1.0}
-                style={styles.textFlex}
-                onPress={() => handleShowPicker(true)}
-              >
-                <Text style={styles.detailText}>{date.getFullYear()}</Text>
-                <Text style={styles.detailText}>{handleMonthName(date)}</Text>
-                <OptionArrow />
-              </TouchableOpacity>
-            </View>
-          ) : (
-            <View style={styles.calendarHeaderView}>
-              <TouchableOpacity
-                activeOpacity={1.0}
-                onPress={() => handleShowPicker(true)}
-              >
-                <Text style={styles.yearText}>{date.getFullYear()}</Text>
-                <Text style={styles.monthText}>{handleMonthName(date)}</Text>
-              </TouchableOpacity>
-              <View style={styles.dropdownContainerView}>
-                <SelectDropdown
-                  data={addPlanData}
-                  dropdownOverlayColor="none"
-                  buttonStyle={styles.dropdownButtonView}
-                  buttonTextStyle={styles.dropdownButtonText}
-                  renderDropdownIcon={(isOpened) => {
-                    return (
-                      <Plus
-                        style={{
-                          transform: [{ rotate: isOpened ? '45deg' : '0deg' }],
-                          marginRight: -10,
-                        }}
-                      />
-                    );
-                  }}
-                  dropdownStyle={styles.dropdownBelowContainerView}
-                  renderCustomizedRowChild={(item) => {
-                    return (
-                      <View style={styles.rowChildView}>
-                        <View
+      <SafeAreaView style={styles.rootContainer}>
+        <View style={marginContainerStyle}>
+          <View style={detail ? undefined : styles.calendarView}>
+            {/* 플랜 생성 페이지 분기처리 */}
+            {detail ? (
+              <View style={styles.calendarHeaderView}>
+                <TouchableOpacity
+                  activeOpacity={1.0}
+                  style={styles.textFlex}
+                  onPress={() => handleShowPicker(true)}
+                >
+                  <Text style={styles.detailText}>{date.getFullYear()}</Text>
+                  <Text style={styles.detailText}>{handleMonthName(date)}</Text>
+                  <OptionArrow />
+                </TouchableOpacity>
+              </View>
+            ) : (
+              <View style={styles.calendarHeaderView}>
+                <TouchableOpacity
+                  activeOpacity={1.0}
+                  onPress={() => handleShowPicker(true)}
+                >
+                  <Text style={styles.yearText}>{date.getFullYear()}</Text>
+                  <Text style={styles.monthText}>{handleMonthName(date)}</Text>
+                </TouchableOpacity>
+                <View style={styles.dropdownContainerView}>
+                  <SelectDropdown
+                    data={addPlanData}
+                    dropdownOverlayColor="none"
+                    buttonStyle={styles.dropdownButtonView}
+                    buttonTextStyle={styles.dropdownButtonText}
+                    renderDropdownIcon={(isOpened) => {
+                      return (
+                        <Plus
                           style={{
-                            ...styles.circleView,
-                            backgroundColor: item[1],
+                            transform: [
+                              { rotate: isOpened ? '45deg' : '0deg' },
+                            ],
+                            marginRight: -10,
                           }}
                         />
-                        <Text>{item[0]}</Text>
-                      </View>
-                    );
-                  }}
-                  rowTextStyle={styles.rowChildText}
-                  rowStyle={styles.rowDividerView}
-                  onSelect={(selectedItem, index) => {
-                    setSelectedFilter(index);
-                  }}
-                  buttonTextAfterSelection={(selectedItem) => {
-                    return selectedItem[0];
-                  }}
-                  rowTextForSelection={(item) => {
-                    return item[0];
-                  }}
-                />
+                      );
+                    }}
+                    dropdownStyle={styles.dropdownBelowContainerView}
+                    renderCustomizedRowChild={(item) => {
+                      return (
+                        <View style={styles.rowChildView}>
+                          <View
+                            style={{
+                              ...styles.circleView,
+                              backgroundColor: item[1],
+                            }}
+                          />
+                          <Text>{item[0]}</Text>
+                        </View>
+                      );
+                    }}
+                    rowTextStyle={styles.rowChildText}
+                    rowStyle={styles.rowDividerView}
+                    onSelect={(selectedItem, index) => {
+                      setSelectedFilter(index);
+                      index === 0
+                        ? handleAddPlan(coupleData.coupleId)
+                        : handleAddPlan(userData.memberId);
+                    }}
+                    buttonTextAfterSelection={(selectedItem) => {
+                      return selectedItem[0];
+                    }}
+                    rowTextForSelection={(item) => {
+                      return item[0];
+                    }}
+                  />
+                </View>
               </View>
-            </View>
-          )}
+            )}
 
-          {/* 캘린더 주간 제목 색상 및 대문자 설정 아직 구현 못함 */}
-          <Calendar
-            monthFormat=""
-            current={formattedDate}
-            key={formattedDate}
-            onDayPress={(day) => {
-              setSelected(day.dateString);
-            }}
-            markingType="multi-dot"
-            hideArrows
-            style={{
-              paddingLeft: 0,
-              paddingRight: 0,
-            }}
-            markedDates={newDaysObject}
-            dayComponent={({ date, state, marking }) => {
-              return (
-                <CheckCalendarDayComponent
-                  detail={detail}
-                  date={date}
-                  state={state}
-                  marking={marking}
-                  setSelected={(date) => handleSelectedDate(date)}
-                  selected={selected}
-                />
-              );
-            }}
-          />
+            {/* 캘린더 주간 제목 색상 및 대문자 설정 아직 구현 못함 */}
+            <Calendar
+              monthFormat=""
+              current={formattedDate}
+              key={formattedDate}
+              onDayPress={(day) => {
+                setSelected(day.dateString);
+              }}
+              markingType="multi-dot"
+              hideArrows
+              style={{
+                paddingLeft: 0,
+                paddingRight: 0,
+              }}
+              markedDates={newDaysObject}
+              dayComponent={({ date, state, marking }) => {
+                return (
+                  <CheckCalendarDayComponent
+                    detail={detail}
+                    date={date}
+                    state={state}
+                    marking={marking}
+                    setSelected={(date) => handleSelectedDate(date)}
+                    selected={selected}
+                  />
+                );
+              }}
+            />
 
-          {detail ? null : (
-            <CheckCalendarDetail
-              selectedMonth={selectedMonth}
-              selectedDay={selectedDay}
-              currentMonth={currentMonth}
-              currentDay={currentDay}
-              myScheduleDetail={myScheduleDetail}
-              partnerScheduleDetail={partnerScheduleDetail}
-              coupleScheduleDetail={coupleScheduleDetail}
-              anniversaryList={anniversaryList}
+            {detail ? null : (
+              <CheckCalendarDetail
+                selectedMonth={selectedMonth}
+                selectedDay={selectedDay}
+                currentMonth={currentMonth}
+                currentDay={currentDay}
+                myScheduleDetail={myScheduleDetail}
+                partnerScheduleDetail={partnerScheduleDetail}
+                coupleScheduleDetail={coupleScheduleDetail}
+                anniversaryList={anniversaryList}
+              />
+            )}
+          </View>
+
+          {show && (
+            <MonthPicker
+              onChange={handleChangeValue}
+              value={date}
+              minimumDate={new Date(1990, 1)}
+              maximumDate={new Date(2025, 5)}
+              locale="en"
             />
           )}
         </View>
-
-        {show && (
-          <MonthPicker
-            onChange={handleChangeValue}
-            value={date}
-            minimumDate={new Date(1990, 1)}
-            maximumDate={new Date(2025, 5)}
-            locale="en"
-          />
-        )}
-      </View>
+      </SafeAreaView>
     </TouchableWithoutFeedback>
   );
 };
@@ -336,6 +356,9 @@ const CheckCalendar = ({
 export default CheckCalendar;
 
 const styles = StyleSheet.create({
+  rootContainer: {
+    flex: 1,
+  },
   marginContainer: {
     paddingTop: 8,
   },
